+++ packages/search-select/src/input.tsx
@@ -91,7 +91,7 @@ export default defineComponent({
     const usingItem: Ref<SelectedItem> = ref(props.defaultUsingItem);
     const menuHoverId = ref('');
     const loading = ref<boolean>(false);
-    const debounceSetMenuList = debounce(300, setMenuList);
+    const debounceSetMenuList = debounce(100, setMenuList);
     // const selectMenuList = ref<ICommonItem[]>([]);
     let isBindEvent = false;
 
@@ -167,7 +167,7 @@ export default defineComponent({
     function documentEnterEvent(e: KeyboardEvent) {
       if (!isBindEvent) return;
       e.preventDefault();
-      const item = menuList.value.find(item => item.id === menuHoverId.value);
+      const item = menuList.value?.find(item => item.id === menuHoverId.value);
       item && handleSelectItem(item);
     }
 
@@ -193,7 +193,9 @@ export default defineComponent({
       showPopover.value = false;
       emit('focus', isFocus.value);
     }
-    function handleInputFocus() {
+    let isOriginFocus = false;
+    function handleInputFocus(event?: FocusEvent) {
+      if (isOriginFocus) return;
       showNoSelectValueError.value = false;
       if (props.mode === SearchInputMode.EDIT && usingItem.value && !isFocus.value) {
         const nodeList = Array.from(
@@ -202,10 +204,15 @@ export default defineComponent({
           ),
         );
         if (!nodeList.length) return;
+        event?.preventDefault();
         const range = document.createRange();
         const selection = window.getSelection();
         range.selectNodeContents(nodeList.at(-1));
         selection?.removeAllRanges();
+        isOriginFocus = true;
+        setTimeout(() => {
+          isOriginFocus = false;
+        }, 200);
         selection.addRange(range); // 注意这里会触发focus事件
         setInputFocus(true, false);
         return;
@@ -254,8 +261,12 @@ export default defineComponent({
           if (
             props.valueBehavior === ValueBehavior.NEED_KEY &&
             menuList.value.some(item => item.id === menuHoverId.value)
-          )
+          ) {
+            if (!usingItem.value && keyword.value?.length) {
+              event.preventDefault();
+            }
             return;
+          }
           handleKeyEnter(event).then(v => v && clearInput());
           break;
         case 'Backspace':
@@ -297,6 +308,7 @@ export default defineComponent({
     function handleKeyBackspace(event: KeyboardEvent) {
       // 删除已选择项
       if (!usingItem.value && !keyword.value) {
+        menuHoverId.value = '';
         emit('delete');
         // eslint-disable-next-line @typescript-eslint/no-misused-promises
         setTimeout(setMenuList, 16);
@@ -562,7 +574,7 @@ export default defineComponent({
       isFocus.value = true;
       showPopover.value = true;
       showNoSelectValueError.value = false;
-      needCursorToEnd && nextTick(setCursorToEnd);
+      !props.getMenuList && needCursorToEnd && nextTick(setCursorToEnd);
       emit('focus', isFocus.value);
     }
     function setSelectedItem(item?: SelectedItem) {
@@ -627,6 +639,11 @@ export default defineComponent({
       usingItem.value.values = [{ id: value, name: value }];
       handleKeyEnter().then(v => v && clearInput());
     }
+    function refleshMenuHover() {
+      if (!usingItem.value) {
+        menuHoverId.value = '';
+      }
+    }
     // expose
     expose({
       inputFocusForWrapper,
@@ -634,6 +651,7 @@ export default defineComponent({
       inputClearForWrapper,
       handleInputFocus,
       isFocus,
+      refleshMenuHover,
     });
 
     return {
@@ -665,6 +683,7 @@ export default defineComponent({
       inputClearForWrapper,
       deleteInputTextNode,
       customPanelSubmit,
+      refleshMenuHover,
       t,
     };
   },